Fix color related issue (2)

+++ Assignment2/src/academy/pocu/comp2500/assignment2/Color.java
@@ -9,8 +9,8 @@ public class Color {
 //        assert(0 <= red && red <= 255);
 //        assert(0 <= green && green <= 255);
 //        assert(0 <= blue && blue <= 255);
-        this.red = Math.max(0, Math.min(red, 255));
-        this.green = Math.max(0, Math.min(green, 255));
-        this.blue = Math.max(0, Math.min(blue, 255));
+        this.red = (short) Math.max(0, Math.min(red, 255));
+        this.green = (short) Math.max(0, Math.min(green, 255));
+        this.blue = (short) Math.max(0, Math.min(blue, 255));
     }
 }
